fix: update direct node_modules paths to new hoisting method of yarn
